Move libs used by recipes into recipes dir
